+++ index.html
@@ -19,5 +19,9 @@
    <body>
      <div id="app"></div>
      <script src="/main.js"></script>
+     <script
+      src="https://kit.fontawesome.com/6478f529f2.js"
+      crossorigin="anonymous"
+    ></script>
    </body>
  </html>
